move warehouse_crafting
